batou/pyproject.toml
[project]
name = "batou"
version = "0.1.0"
description = "MCP server for structured entity operations on Context Lake workspaces"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0.0",
]

[project.scripts]
batou = "batou.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/batou"]
